+++ minishell/src/executions/ms_execmd.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/12 10:21:17 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/06/12 11:13:58 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/06/17 18:04:12 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
